refactor: adjustments files
