+++ access-events-app/src/Components/RenderIf/index.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { FunctionComponent } from "react";
 import { RenderIfProps } from "./RenderIf.types";
 
@@ -5,7 +6,11 @@ const RenderIf: FunctionComponent<RenderIfProps> = ({
   condition,
   children,
 }) => {
-  return condition ? <>{children}</> : null;
+  if (!condition) {
+    return null;
+  }
+
+  return <React.Fragment>{children}</React.Fragment>;
 };
 
 export default RenderIf;
