+++ access-events-app/src/contexts/ToastContext/Toast.context.ts
@@ -1,0 +1,14 @@
+import {
+  ToastPosition,
+  ToastVariant,
+} from "../../Components/ToastComponent/Toast.component.types";
+
+export interface Toast {
+  message: string;
+  variant: ToastVariant;
+  position: ToastPosition;
+}
+
+export interface ToastContextType {
+  toast: (options: Toast) => void;
+}
