+++ access-events-app/src/modules/Home/Components/SuggestionSection/index.tsx
@@ -1,5 +1,5 @@
 import { FunctionComponent } from "react";
-import { BoxContainer } from "../../../../Components/BoxContainer";
+import { BoxContainer } from "../../../../Components/BoxContainer/BoxContainer.component";
 import useStyles from "./styles";
 import ModalNewActivities from "../Modal/SolicitationModal";
 import BoxOptionsList from "./Components/BoxOptionsList";
